requirements.txt
anyio==3.6.2
fastapi==0.89.1
idna==3.4
pydantic==1.10.4
python-dotenv==0.21.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
